requirements.txt
Django==4.2.5
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
djangorestframework==3.14.0
hashids==1.3.1
google-api-python-client==1.7.9
google-auth-oauthlib==0.4.0
mysqlclient
